# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "FilePolicy": {
        "properties": {
          "audios": {
            "type": "boolean"
          },
          "documents": {
            "type": "boolean"
          },
          "images": {
            "type": "boolean"
          },
          "plainText": {
            "type": "boolean"
          }
        },
        "required": [
          "plainText",
          "documents",
          "images",
          "audios"
        ],
        "type": "object"
      },
      "Filter": {
        "properties": {
          "advancedQuery": {
            "type": "string"
          },
          "mode": {
            "enum": [
              "simple",
              "advanced"
            ],
            "type": "string"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/FilterRule"
            },
            "type": "array"
          }
        },
        "required": [
          "mode",
          "rules",
          "advancedQuery"
        ],
        "type": "object"
      },
...
      "Scope": {
        "properties": {
          "rule": {
            "enum": [
              "combined",
              "shared",
              "per_tag",
              "all_combinations",
              "custom"
            ],
            "type": "string"
          },
          "scopes": {
            "items": {
              "items": {
                "$ref": "#/components/schemas/TagRef"
              },
              "type": "array"
            },
            "type": "array"
          }
        },
        "required": [
          "rule",
          "scopes"
        ],
        "type": "object"
      },
...
      "TagRef": {
        "properties": {
          "kind": {
            "enum": [
              "literal",
              "dynamic"
            ],
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "kind",
          "value"
        ],
        "type": "object"
      },
      "apiError": {
        "properties": {
          "code": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "fields": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          }
        },
        "required": [
          "error",
          "code"
        ],
        "type": "object"
      },
...
      "taskDTO": {
        "properties": {
          "configRevision": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string"
          },
          "cronExpression": {
            "type": "string"
          },
          "cronTimezone": {
            "type": "string"
          },
          "customMetadata": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "customTags": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "destinationBankId": {
            "type": "string"
          },
          "destinationCredentialId": {
            "type": "string"
          },
          "destinationLocked": {
            "type": "boolean"
          },
          "enabled": {
            "type": "boolean"
          },
          "filePolicy": {
            "$ref": "#/components/schemas/FilePolicy"
          },
          "filePolicyMode": {
            "enum": [
              "global",
              "override"
            ],
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "inlineMultimodalEnabled": {
            "type": "boolean"
          },
          "inlineMultimodalMode": {
            "enum": [
              "global",
              "override"
            ],
            "type": "string"
          },
          "itemCount": {
            "type": "integer"
          },
          "lastRun": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/runDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "type": "string"
          },
          "nextRunAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "observationScope": {
            "$ref": "#/components/schemas/Scope"
          },
          "observationScopeMode": {
            "enum": [
              "global",
              "override"
            ],
            "type": "string"
          },
          "policyRevision": {
            "type": "integer"
          },
          "reconcileRequired": {
            "type": "boolean"
          },
          "retainStrategy": {
            "type": "string"
          },
          "running": {
            "type": "boolean"
          },
          "sourceConfig": {
            "additionalProperties": {},
            "type": "object"
          },
          "sourceCredentialId": {
            "type": "string"
          },
          "sourceFilter": {
            "$ref": "#/components/schemas/Filter"
          },
          "sourceType": {
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/taskStateDTO"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "observationScopeMode",
          "observationScope",
          "id",
          "name",
          "enabled",
          "sourceType",
          "sourceCredentialId",
          "sourceConfig",
          "sourceFilter",
          "destinationCredentialId",
          "destinationBankId",
          "retainStrategy",
          "customTags",
          "customMetadata",
          "filePolicyMode",
          "filePolicy",
          "inlineMultimodalMode",
          "inlineMultimodalEnabled",
          "cronExpression",
          "cronTimezone",
          "configRevision",
          "policyRevision",
          "reconcileRequired",
          "destinationLocked",
          "running",
          "nextRunAt",
          "lastRun",
          "itemCount",
          "state",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "scheme": "basic",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Hindsight Ingestion API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/tasks/{id}": {
      "delete": {
        "operationId": "delete_api_tasks_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "deleteDocuments",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Basic Auth required"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiError"
                }
              }
            },
            "description": "API error"
          }
        }
      },
      "get": {
        "operationId": "get_api_tasks_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/taskDTO"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Basic Auth required"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiError"
                }
              }
            },
            "description": "API error"
          }
        }
      },
      "patch": {
        "operationId": "patch_api_tasks_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "cronExpression": {
                    "type": "string"
                  },
                  "cronTimezone": {
                    "type": "string"
                  },
                  "customMetadata": {
                    "additionalProperties": {
                      "type": "string"
                    },
                    "type": "object"
                  },
                  "customTags": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "destinationBankId": {
                    "type": "string"
                  },
                  "destinationCredentialId": {
                    "type": "string"
                  },
                  "enabled": {
                    "type": "boolean"
                  },
                  "filePolicy": {
                    "$ref": "#/components/schemas/FilePolicy"
                  },
                  "filePolicyMode": {
                    "enum": [
                      "global",
                      "override"
                    ],
                    "type": "string"
                  },
                  "inlineMultimodalEnabled": {
                    "type": "boolean"
                  },
                  "inlineMultimodalMode": {
                    "enum": [
                      "global",
                      "override"
                    ],
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "observationScope": {
                    "$ref": "#/components/schemas/Scope"
                  },
                  "observationScopeMode": {
                    "enum": [
                      "global",
                      "override"
                    ],
                    "type": "string"
                  },
                  "retainStrategy": {
                    "type": "string"
                  },
                  "sourceConfig": {
                    "additionalProperties": {},
                    "type": "object"
                  },
                  "sourceCredentialId": {
                    "type": "string"
                  },
                  "sourceFilter": {
                    "$ref": "#/components/schemas/Filter"
                  },
                  "sourceType": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/taskDTO"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Basic Auth required"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiError"
                }
              }
            },
            "description": "API error"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "basicAuth": []
    }
  ]
}